refactor: add useLabel controller
