+++ packages/components/src/components/input-date-picker/input-date-picker.tsx
@@ -35,8 +35,9 @@ import {
   reposition,
 } from "../../utils/floating-ui";
 import { numberKeys } from "../../utils/key";
-import { connectLabel, disconnectLabel, LabelableComponent, getLabelText } from "../../utils/label";
+import { getLabelText } from "../../utils/label";
 import { getIconScale } from "../../utils/component";
+import { type LabelableComponent, useLabel } from "../../controllers/useLabel";
 import {
   getDateFormatSupportedLocale,
   getSupportedNumberingSystem,
@@ -142,6 +143,8 @@ export class InputDatePicker extends LitElement implements FloatingUIComponent, 
   })(this);
 
   labelEl?: Label["el"];
+
+  labelable = useLabel<this>()(this);
 
   transitionProp = "opacity" as const;
 
@@ -416,7 +419,6 @@ export class InputDatePicker extends LitElement implements FloatingUIComponent, 
       this.openHandler();
     }
 
-    connectLabel(this);
     this.setFilteredPlacements();
     connectFloatingUI(this);
   }
@@ -500,7 +502,6 @@ export class InputDatePicker extends LitElement implements FloatingUIComponent, 
   }
 
   override disconnectedCallback(): void {
-    disconnectLabel(this);
     disconnectFloatingUI(this);
   }
