+++ packages/components/src/components/rating/rating.tsx
@@ -9,7 +9,8 @@ import {
   stringOrBoolean,
 } from "@arcgis/lumina";
 import { guid } from "../../utils/guid";
-import { connectLabel, disconnectLabel, LabelableComponent, getLabelText } from "../../utils/label";
+import { getLabelText } from "../../utils/label";
+import { type LabelableComponent, useLabel } from "../../controllers/useLabel";
 import { Scale, Status } from "../interfaces";
 import { InternalLabel } from "../functional/InternalLabel";
 import { Validation } from "../functional/Validation";
@@ -80,6 +81,8 @@ export class Rating extends LitElement implements LabelableComponent {
   private focusSetter = useSetFocus<this>()(this);
 
   private interactiveContainer = useInteractive(this);
+
+  labelable = useLabel<this>()(this);
 
   //#endregion
 
@@ -197,10 +200,6 @@ export class Rating extends LitElement implements LabelableComponent {
     this.listen("pointerover", this.handleRatingPointerOver);
   }
 
-  override connectedCallback(): void {
-    connectLabel(this);
-  }
-
   async load(): Promise<void> {
     this.requestUpdate("value");
   }
@@ -237,10 +236,6 @@ export class Rating extends LitElement implements LabelableComponent {
 
   loaded(): void {
     this.labelElements = Array.from(this.renderRoot.querySelectorAll("label"));
-  }
-
-  override disconnectedCallback(): void {
-    disconnectLabel(this);
   }
 
   //#endregion
