+++ packages/components/src/components/segmented-control/segmented-control.tsx
@@ -12,12 +12,8 @@ import {
 } from "@arcgis/lumina";
 import { useDirection } from "@arcgis/lumina/controllers";
 import { slotChangeGetAssignedElements } from "../../utils/dom";
-import {
-  connectLabel,
-  disconnectLabel,
-  type LabelableComponent,
-  getLabelText,
-} from "../../utils/label";
+import { getLabelText } from "../../utils/label";
+import { type LabelableComponent, useLabel } from "../../controllers/useLabel";
 import type { Appearance, Layout, Scale, Status, Width } from "../interfaces";
 import { InternalLabel } from "../functional/InternalLabel";
 import { Validation } from "../functional/Validation";
@@ -73,6 +69,8 @@ export class SegmentedControl extends LitElement implements LabelableComponent {
   private focusSetter = useSetFocus<this>()(this);
 
   private interactiveContainer = useInteractive(this);
+
+  labelable = useLabel<this>()(this);
 
   //#endregion
 
@@ -181,10 +179,6 @@ export class SegmentedControl extends LitElement implements LabelableComponent {
     this.listen("click", this.handleClick);
   }
 
-  override connectedCallback(): void {
-    connectLabel(this);
-  }
-
   override willUpdate(changes: PropertyValues<this>): void {
     /* TODO: [MIGRATION] First time Lit calls willUpdate(), changes will include not just properties provided by the user, but also any default values your component set.
     To account for this semantics change, the checks for (this.hasUpdated || value != defaultValue) was added in this method
@@ -209,10 +203,6 @@ export class SegmentedControl extends LitElement implements LabelableComponent {
 
   loaded(): void {
     this.formSupport.overrideDefaultValue(this.value);
-  }
-
-  override disconnectedCallback(): void {
-    disconnectLabel(this);
   }
 
   //#endregion
